backend/requirements.txt
fastapi
uvicorn[standard]
google-api-python-client
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-storage
google-cloud-secret-manager
psycopg2-binary
langchain
langchain-google-vertexai
python-docx
PyPDF2
redis
celery
pydantic-settings
pytesseract
opencv-python-headless
